ai_agents/geo_location/requirements.txt
torch>=2.0.0
transformers>=4.35.0
tqdm>=4.64.0
python-dotenv>=0.21.0
huggingface_hub>=0.17.0
accelerate>=0.20.0
google-generativeai>=0.3.0
